requirements.txt
opencv-python==4.5.3.56
numpy==1.21.1
pyserial==3.5
pillow==8.3.1
argparse==1.4.0